+++ trace/driver_gtrace.go
@@ -454,6 +454,25 @@ func (t *Driver) Compose(x *Driver, opts ...DriverComposeOption) *Driver {
 		}
 	}
 	{
+		h1 := t.OnConnStreamFinish
+		h2 := x.OnConnStreamFinish
+		ret.OnConnStreamFinish = func(info DriverConnStreamFinishInfo) {
+			if options.panicCallback != nil {
+				defer func() {
+					if e := recover(); e != nil {
+						options.panicCallback(e)
+					}
+				}()
+			}
+			if h1 != nil {
+				h1(info)
+			}
+			if h2 != nil {
+				h2(info)
+			}
+		}
+	}
+	{
 		h1 := t.OnConnDial
 		h2 := x.OnConnDial
 		ret.OnConnDial = func(d DriverConnDialStartInfo) func(DriverConnDialDoneInfo) {
@@ -1055,6 +1074,13 @@ func (t *Driver) onConnStreamCloseSend(d DriverConnStreamCloseSendStartInfo) fun
 	}
 	return res
 }
+func (t *Driver) onConnStreamFinish(info DriverConnStreamFinishInfo) {
+	fn := t.OnConnStreamFinish
+	if fn == nil {
+		return
+	}
+	fn(info)
+}
 func (t *Driver) onConnDial(d DriverConnDialStartInfo) func(DriverConnDialDoneInfo) {
 	fn := t.OnConnDial
 	if fn == nil {
@@ -1235,6 +1261,7 @@ func (t *Driver) onGetCredentials(d DriverGetCredentialsStartInfo) func(DriverGe
 	}
 	return res
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnInit(t *Driver, c *context.Context, call call, endpoint string, database string, secure bool) func(error) {
 	var p DriverInitStartInfo
@@ -1250,6 +1277,7 @@ func DriverOnInit(t *Driver, c *context.Context, call call, endpoint string, dat
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnWith(t *Driver, c *context.Context, call call, endpoint string, database string, secure bool) func(error) {
 	var p DriverWithStartInfo
@@ -1265,6 +1293,7 @@ func DriverOnWith(t *Driver, c *context.Context, call call, endpoint string, dat
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnClose(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverCloseStartInfo
@@ -1277,6 +1306,7 @@ func DriverOnClose(t *Driver, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnPoolNew(t *Driver, c *context.Context, call call) func() {
 	var p DriverConnPoolNewStartInfo
@@ -1288,6 +1318,7 @@ func DriverOnPoolNew(t *Driver, c *context.Context, call call) func() {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnPoolRelease(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverConnPoolReleaseStartInfo
@@ -1300,6 +1331,7 @@ func DriverOnPoolRelease(t *Driver, c *context.Context, call call) func(error) {
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnResolve(t *Driver, call call, target string, resolved []string) func(error) {
 	var p DriverResolveStartInfo
@@ -1313,6 +1345,7 @@ func DriverOnResolve(t *Driver, call call, target string, resolved []string) fun
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnStateChange(t *Driver, c *context.Context, call call, endpoint EndpointInfo, state ConnState) func(state ConnState) {
 	var p DriverConnStateChangeStartInfo
@@ -1327,6 +1360,7 @@ func DriverOnConnStateChange(t *Driver, c *context.Context, call call, endpoint 
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnInvoke(t *Driver, c *context.Context, call call, endpoint EndpointInfo, m Method) func(_ error, issues []Issue, opID string, state ConnState, metadata map[string][]string) {
 	var p DriverConnInvokeStartInfo
@@ -1345,6 +1379,7 @@ func DriverOnConnInvoke(t *Driver, c *context.Context, call call, endpoint Endpo
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnNewStream(t *Driver, c *context.Context, call call, endpoint EndpointInfo, m Method) func(_ error, state ConnState) {
 	var p DriverConnNewStreamStartInfo
@@ -1360,6 +1395,7 @@ func DriverOnConnNewStream(t *Driver, c *context.Context, call call, endpoint En
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnStreamRecvMsg(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverConnStreamRecvMsgStartInfo
@@ -1372,6 +1408,7 @@ func DriverOnConnStreamRecvMsg(t *Driver, c *context.Context, call call) func(er
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnStreamSendMsg(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverConnStreamSendMsgStartInfo
@@ -1384,6 +1421,7 @@ func DriverOnConnStreamSendMsg(t *Driver, c *context.Context, call call) func(er
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnStreamCloseSend(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverConnStreamCloseSendStartInfo
@@ -1396,6 +1434,16 @@ func DriverOnConnStreamCloseSend(t *Driver, c *context.Context, call call) func(
 		res(p)
 	}
 }
+
+// Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
+func DriverOnConnStreamFinish(t *Driver, c context.Context, call call, e error) {
+	var p DriverConnStreamFinishInfo
+	p.Context = c
+	p.Call = call
+	p.Error = e
+	t.onConnStreamFinish(p)
+}
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnDial(t *Driver, c *context.Context, call call, endpoint EndpointInfo) func(error) {
 	var p DriverConnDialStartInfo
@@ -1409,6 +1457,7 @@ func DriverOnConnDial(t *Driver, c *context.Context, call call, endpoint Endpoin
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnBan(t *Driver, c *context.Context, call call, endpoint EndpointInfo, state ConnState, cause error) func(state ConnState) {
 	var p DriverConnBanStartInfo
@@ -1424,6 +1473,7 @@ func DriverOnConnBan(t *Driver, c *context.Context, call call, endpoint Endpoint
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnAllow(t *Driver, c *context.Context, call call, endpoint EndpointInfo, state ConnState) func(state ConnState) {
 	var p DriverConnAllowStartInfo
@@ -1438,6 +1488,7 @@ func DriverOnConnAllow(t *Driver, c *context.Context, call call, endpoint Endpoi
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnPark(t *Driver, c *context.Context, call call, endpoint EndpointInfo) func(error) {
 	var p DriverConnParkStartInfo
@@ -1451,6 +1502,7 @@ func DriverOnConnPark(t *Driver, c *context.Context, call call, endpoint Endpoin
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnConnClose(t *Driver, c *context.Context, call call, endpoint EndpointInfo) func(error) {
 	var p DriverConnCloseStartInfo
@@ -1464,6 +1516,7 @@ func DriverOnConnClose(t *Driver, c *context.Context, call call, endpoint Endpoi
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnRepeaterWakeUp(t *Driver, c *context.Context, call call, name string, event string) func(error) {
 	var p DriverRepeaterWakeUpStartInfo
@@ -1478,6 +1531,7 @@ func DriverOnRepeaterWakeUp(t *Driver, c *context.Context, call call, name strin
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnBalancerInit(t *Driver, c *context.Context, call call, name string) func(error) {
 	var p DriverBalancerInitStartInfo
@@ -1491,6 +1545,7 @@ func DriverOnBalancerInit(t *Driver, c *context.Context, call call, name string)
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnBalancerClose(t *Driver, c *context.Context, call call) func(error) {
 	var p DriverBalancerCloseStartInfo
@@ -1503,6 +1558,7 @@ func DriverOnBalancerClose(t *Driver, c *context.Context, call call) func(error)
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnBalancerChooseEndpoint(t *Driver, c *context.Context, call call) func(endpoint EndpointInfo, _ error) {
 	var p DriverBalancerChooseEndpointStartInfo
@@ -1516,6 +1572,7 @@ func DriverOnBalancerChooseEndpoint(t *Driver, c *context.Context, call call) fu
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnBalancerClusterDiscoveryAttempt(t *Driver, c *context.Context, call call, address string) func(error) {
 	var p DriverBalancerClusterDiscoveryAttemptStartInfo
@@ -1529,6 +1586,7 @@ func DriverOnBalancerClusterDiscoveryAttempt(t *Driver, c *context.Context, call
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnBalancerUpdate(t *Driver, c *context.Context, call call, needLocalDC bool) func(endpoints []EndpointInfo, added []EndpointInfo, dropped []EndpointInfo, localDC string) {
 	var p DriverBalancerUpdateStartInfo
@@ -1545,6 +1603,7 @@ func DriverOnBalancerUpdate(t *Driver, c *context.Context, call call, needLocalD
 		res(p)
 	}
 }
+
 // Internals: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#internals
 func DriverOnGetCredentials(t *Driver, c *context.Context, call call) func(token string, _ error) {
 	var p DriverGetCredentialsStartInfo