+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added `trace.DriverConnStreamEvents` details bit
+* Added `trace.Driver.OnConnStreamFinish` event
 
 ## v3.66.1
 * Added flush messages from buffer before close topic writer