+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.66.2
 * Added `trace.DriverConnStreamEvents` details bit
 * Added `trace.Driver.OnConnStreamFinish` event